vqa_project/requirements.txt
torch==2.1.0
transformers==4.39.0
datasets==2.5.0
Pillow==9.2.0
tqdm==4.64.1
numpy==1.23.1
scikit-learn==1.1.1
